+++ src/app/error.tsx
@@ -1,5 +1,5 @@
 'use client';
-import ErrorOrNotFound from '@/components/common/ErrorOrNotFound/ErrorOrNotFound';
+import ErrorOrNotFound from '@/components/common/ErrorOrNotFound';
 import { useEffect } from 'react';
 
 type Props = {
