+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import ErrorOrNotFound from '@/components/common/ErrorOrNotFound/ErrorOrNotFound';
+import ErrorOrNotFound from '@/components/common/ErrorOrNotFound';
 
 export default function GlobalNotFound() {
   return <ErrorOrNotFound text='페이지를 찾을 수 없습니다.' />;
