+++ src/components/common/ErrorOrNotFound/index.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import styles from './css/ErrorOrNotFound.module.css';
+import styles from './css/index.module.css';
 
 type Props = {
   text: string;
